requirements.txt
colorlog==6.9.0
pyyaml==6.0.2
